pyproject.toml
[tool.poetry]
name = "krankaoke_backend"
version = "0.1.1"
description = ""
authors = ["Kees van Ekeren <kg.v.ekeren@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.3"
djangorestframework = "^3.11.0"
pendulum = "^2.0.5"
factory_boy = "^2.12.0"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.4"
drf-yasg = "^1.17.1"
django-cors-headers = "^3.2.1"
mypy = "^0.770"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
black = "^19.10b0"
isort = "^4.3.21"
coverage = "^5.0.3"
ipython = "^7.12.0"
ipdb = "^0.12.3"
django-extensions = "^2.2.8"
pytest-cov = "^2.8.1"
pytest-django = "^3.8.0"

[tool.black]
line_length = 88

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
skip_glob = "**/migrations/*,node_modules"

[tool.pytest]
addopts = "-v --cov --ignore=node_modules --cov=krankaoke_backend"
DJANGO_SETTINGS_MODULE = "krankaoke_backend.settings"
python_files = "tests.py test_*.py *_tests.py"

[tool.coverage.run]
source = "krankaoke_backend"
branch = true
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

